Changes by Mizael

+++ src/main/java/inatel/br/cdg/Inimigo/Inimigo.java
@@ -2,10 +2,10 @@ package inatel.br.cdg.Inimigo;
 
 public abstract class Inimigo {
 
-    double vida;
-    double ataque;
+    int vida;
+    int ataque;
 
-    public Inimigo(double vida, double ataque) {
+    public Inimigo(int vida, int ataque) {
         this.vida = vida;
         this.ataque = ataque;
     }
